Fix test and move them over to a dedicated file
these tests are not bound to the cmdline
